Move internal modules to core
