+++ core/src/main/scala/hedgehog/core/Tree.scala
@@ -1,4 +1,4 @@
-package hedgehog
+package hedgehog.core
 
 import hedgehog.predef._
